requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
cloudinary==1.32.0
coverage==7.2.2
dj-database-url==0.5.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gunicorn==20.1.0
iniconfig==2.0.0
oauthlib==3.2.2
pathspec==0.11.1
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.6.0
pytest==7.2.2
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.3
